criu: Add lib-bsd
Provides strlcpy(), strlcat() and optionally setproctitle_init(),
setproctitle().

Get them from loaded in runtime libbsd.so or use home-made criu
versions.
