+++ criu/Makefile.crtools
@@ -38,6 +38,7 @@ obj-y			+= irmap.o
 obj-y			+= kcmp-ids.o
 obj-y			+= kerndat.o
 obj-y			+= libnetlink.o
+obj-y			+= lib-bsd.o
 obj-y			+= log.o
 obj-y			+= lsm.o
 obj-y			+= mem.o
@@ -73,7 +74,6 @@ obj-y			+= sk-tcp.o
 obj-y			+= sk-unix.o
 obj-y			+= sockets.o
 obj-y			+= stats.o
-obj-y			+= string.o
 obj-y			+= sysctl.o
 obj-y			+= sysfs_parse.o
 obj-y			+= timerfd.o
